requirements.txt
gevent-websocket
python-socketio
bottle
bcrypt
python-dotenv
PyJWT
